+++ R/create_SE.R
@@ -120,11 +120,7 @@ create_SE <- function(df_,
   
   ## Combine all controls with respective treatments.
   # Merge raw data back with groupings.
-  dfs <- merge(
-    df_, 
-    mapping_entries, 
-    by = c(colnames(rowdata), colnames(coldata))
-  )
+  dfs <- df_[mapping_entries, on = c(colnames(rowdata), colnames(coldata))]
 
   # Remove all rowdata and coldata. 
   groupings <- dfs$groupings